chore(PlayerSwap): more accurate swap cooldown time

+++ src/main/java/io/github/mal32/endergames/game/game/PlayerSwapManager.java
@@ -16,7 +16,7 @@ import org.bukkit.potion.PotionEffectType;
 public class PlayerSwapManager extends AbstractTask {
   private static final int SWAP_COOLDOWN_SECONDS = 60 * 3;
   private static final int CLOCK_SPEED_SECONDS = 5;
-  private int swapCooldown = SWAP_COOLDOWN_SECONDS;
+  private double swapCooldown = SWAP_COOLDOWN_SECONDS;
 
   public PlayerSwapManager(EnderGames plugin) {
     super(plugin);
